refactor(workspace): split 'superdesk-worspace' and 'content' modules.
`superdesk-workspace` becomes:
```
.
├── directives
│   ├── EditWorkspaceDirective.js
│   ├── WorkspaceDropdownDirective.js
│   ├── WorkspaceSidenavDirective.js
│   └── index.js
├── services
│   ├── WorkspaceService.js
│   └── index.js
├── tests
│   └── workspace.spec.js
├── views
│   ├── edit-workspace-modal.html
│   ├── workspace-dropdown.html
│   ├── workspace-sidenav-items.html
│   └── workspace-sidenav.html
└── index.js

11 directories, 14 files
```

`superdesk-workspace/content` becomes:
```
.
├── controllers
│   ├── ContentProfilesController.js
│   └── index.js
├── directives
│   ├── ContentCreateDirective.js
│   ├── ContentProfileSchemaEditor.js
│   ├── ItemProfileDirective.js
│   ├── StringToArrayDirective.js
│   └── index.js
├── services
│   ├── ContentService.js
│   └── index.js
├── styles
│   └── profiles.less
├── tests
│   └── content.spec.js
├── views
│   ├── profile-settings.html
│   ├── schema-editor.html
│   └── sd-content-create.html
├── constants.js
└── index.js

6 directories, 16 files
```
