+++ scripts/superdesk-workspace/content/controllers/index.js
@@ -1,0 +1,1 @@
+export { ContentProfilesController } from './ContentProfilesController';
