+++ scripts/superdesk-workspace/content/directives/ItemProfileDirective.js
@@ -1,0 +1,14 @@
+ItemProfileDirective.$inject = ['content'];
+export function ItemProfileDirective(content) {
+    return {
+        scope: {profileId: '=profile'},
+        template: '{{ profile }}',
+        link: function(scope) {
+            content.getTypesLookup().then(function(lookup) {
+                scope.profile = lookup[scope.profileId] ?
+                    lookup[scope.profileId].label :
+                    scope.profileId;
+            });
+        }
+    };
+}
