+++ scripts/superdesk-workspace/content/directives/index.js
@@ -1,0 +1,4 @@
+export { ContentCreateDirective } from './ContentCreateDirective';
+export { ContentProfileSchemaEditor } from './ContentProfileSchemaEditor';
+export { StringToArrayDirective } from './StringToArrayDirective';
+export { ItemProfileDirective } from './ItemProfileDirective';
