+++ scripts/superdesk-workspace/content/services/index.js
@@ -1,0 +1,1 @@
+export { ContentService } from './ContentService';
