+++ scripts/superdesk-workspace/content/tests/content.spec.js
