+++ scripts/superdesk-workspace/directives/index.js
@@ -1,0 +1,3 @@
+export { WorkspaceDropdownDirective } from './WorkspaceDropdownDirective';
+export { WorkspaceSidenavDirective } from './WorkspaceSidenavDirective';
+export { EditWorkspaceDirective } from './EditWorkspaceDirective';
