+++ scripts/superdesk-workspace/services/index.js
@@ -1,0 +1,1 @@
+export { WorkspaceService } from './WorkspaceService';
