+++ scripts/superdesk-workspace/tests/workspace.spec.js
